requirements.txt
sqlalchemy ==1.4.3
pytest-datadir ==1.3.1
setuptools ==56.0.0
twine ==3.4.1
wheel ==0.36.2
